+++ src/App.tsx
@@ -29,8 +29,8 @@ import TimetablePage from 'pages/TimetablePage/MainTimetablePage';
 import CafeteriaPage from 'pages/Cafeteria';
 import MetaHelmet from 'components/MetaHelmet';
 import ModifyInfoPage from 'pages/Auth/ModifyInfoPage';
-import AddReviewPage from 'pages/StoreReviewPage/AddReviewPage';
-import EditReviewPage from 'pages/StoreReviewPage/EditReviewPage';
+import AddReviewPage from 'pages/Store/StoreReviewPage/AddReviewPage';
+import EditReviewPage from 'pages/Store/StoreReviewPage/EditReviewPage';
 import ReviewReportingPage from 'pages/Store/StoreDetailPage/Review/components/ReviewReporting';
 import ModifyTimetablePage from 'pages/TimetablePage/ModifyTimetablePage';
 import GraduationCalculatorPage from 'pages/GraduationCalculatorPage';
